feat(mvp-11): ajout de la creation du profile

+++ src/tests/setup.ts
@@ -1,5 +1,8 @@
+import dotenv from 'dotenv';
 import { MongoMemoryServer } from 'mongodb-memory-server';
 import mongoose from 'mongoose';
+
+dotenv.config({ path: '.env.test' });
 
 let mongo: MongoMemoryServer;
 
@@ -10,18 +13,25 @@ beforeAll(async () => {
 
   const uri = mongo.getUri();
   await mongoose.connect(uri);
+
+  console.log(`🧪 MongoDB en mémoire lancée sur ${uri}`);
+});
+
+afterEach(async () => {
+  const db = mongoose.connection.db;
+  if (!db) {
+    console.warn('⚠️ Aucune connexion DB disponible dans afterEach');
+    return;
+  }
+
+  const collections = await db.collections();
+  for (const collection of collections) {
+    await collection.deleteMany({});
+  }
 });
 
 afterAll(async () => {
   await mongoose.disconnect();
   await mongo.stop();
-});
-
-afterEach(async () => {
-  const db = mongoose.connection.db;
-  const collections = await db!.collections();
-
-  for (const collection of collections) {
-    await collection.deleteMany({});
-  }
+  console.log('🧹 MongoMemoryServer arrêté proprement');
 });
